% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri_target_graph.R
\name{AssignTargetCluster}
\alias{AssignTargetCluster}
\title{Try to find optimal cluster number fro fuzzy k meansand make plots for k selection}
\usage{
AssignTargetCluster(
  normalize_matrix,
  select.k = 7,
  group.label,
  p.cutoff = 0.05
)
}
\arguments{
\item{normalize_matrix}{A matrix from Run_LINE()}

\item{select.k}{A numeric for k numbers to use}

\item{group.label}{A factor vector subtype label for the samples}

\item{p.cutoff}{A numeric variable indicating the p-value cutoff when assigning clusters to groups}
}
\value{
A list in which GeneList shows targets genes to each group, Label stores the group assigned to each cluster
}
\description{
Try to find optimal cluster number fro fuzzy k meansand make plots for k selection
}
