% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildCorList}
\alias{BuildCorList}
\title{Generate a list of correlation matrices for later analysis}
\usage{
BuildCorList(mrna.mat, mirna.mat, group.label, numCores = 1)
}
\arguments{
\item{mrna.mat}{A matrix of mRNA expression (genes at rows and samples at columns)}

\item{mirna.mat}{A matrix of miRNA expression (genes at rows and samples at columns)}

\item{group.label}{A factor vector subtype label for the samples}

\item{numCores}{An integer indicating the number of cores to use}
}
\description{
Generate a list of correlation matrices for later analysis
}
