% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildTargetMat}
\alias{BuildTargetMat}
\title{Generate a binarized target matrix}
\usage{
BuildTargetMat(
  mrna.mat,
  mirna.mat,
  DB,
  org = "hsa",
  gene_id = "target_symbol",
  min.cutoff = 1
)
}
\arguments{
\item{mrna.mat}{*A matrix of mRNA expression (genes at rows and samples at columns)}

\item{mirna.mat}{*A matrix of miRNA expression (genes at rows and samples at columns)}

\item{DB}{*A string vector showing which databases to use}

\item{org}{*A string indiciating whether human or mice target data use}

\item{gene_id}{*A character string showing which type of gene ID to use. Can be: target_symbol,target_entrez or target_ensembl}

\item{min.cutoff}{An integer indicating minimum database predictions required for target information}
}
\value{
A binary target matrix. 1 means predicted target pair, and 0 means not.
}
\description{
Generate a binarized target matrix
}
