% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default.pal}
\alias{default.pal}
\title{Generate a vector of colors to use}
\usage{
default.pal(n, set)
}
\arguments{
\item{n}{Number of levels}

\item{set}{Pallete to choose}
}
\description{
Generate a vector of colors to use
}
