% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri_target_maxdES.R
\name{.Calc_NES_bkgd}
\alias{.Calc_NES_bkgd}
\title{Calculating background ES distribution for normalizing dES}
\usage{
.Calc_NES_bkgd(gs, rk, direction, n.iter, alpha)
}
\arguments{
\item{gs}{A string vector for set of targets of interests. Could be an element from output of GetLE()}

\item{rk}{A numeric vector for all mRNAs ranked by correlation}

\item{direction}{A character for the direction of enrichment}

\item{n.iter}{An integer for the number of background to generate. Default is 500}

\item{alpha}{A numeric value for the gsea param}
}
\value{
A numeric vector of background normalized ES distribution
}
\description{
Calculating background ES distribution for normalizing dES
}
