% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri_target_maxdES.R
\name{.Diff_by_Remove}
\alias{.Diff_by_Remove}
\title{Remove each gene in the gs and see what is the changes in the dES}
\usage{
.Diff_by_Remove(gs, main.rk, ref.rk, direction, n.iter, alpha)
}
\arguments{
\item{gs}{A character vector for a set of targets. One element from output of GetLE().}

\item{main.rk}{A numeric vector for all mRNAs ranked by correlation in the group of interest}

\item{ref.rk}{A numeric vector for all mRNAs ranked by correlation in the group to compare with}

\item{direction}{A character for the direction of enrichment}

\item{n.iter}{An integer for the number of background to generate. Default is 500}

\item{alpha}{A numeric value for the gsea param}
}
\value{
A numberic vector of the dES for removing each gene in gs
}
\description{
Remove each gene in the gs and see what is the changes in the dES
}
