% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffreg.R
\name{.diffreg_ks}
\alias{.diffreg_ks}
\title{Run two-group differential enrichment using KS test
#'}
\usage{
.diffreg_ks(
  target.mat,
  cor.list,
  min.target.number = 5,
  direction = "Negative",
  alpha
)
}
\arguments{
\item{target.mat}{A binary matrix indicating target with miRNA on columns and mRNA on rows}

\item{cor.list}{A list of correlation matrices with miRNA on columns and mRNA on rows}

\item{min.target.number}{A integer indicating minimum number of targets requried for doing diffreg}

\item{direction}{A character string specifying the direction of enrichment, must be "Negative" or "Positive"}

\item{alpha}{Weight for running AD/KS test}
}
\value{
A data frame with ES and test results for each miRNA
}
\description{
Run two-group differential enrichment using KS test
#'
}
