% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri_target_maxdES.R
\name{Find_Target_Max_dES}
\alias{Find_Target_Max_dES}
\title{Do backward random search to find set of targets that maximize the normalized dES}
\usage{
Find_Target_Max_dES(
  gs,
  cor.list,
  main,
  ref,
  miR,
  direction = c("Negative", "Positive"),
  min.gs.size = 20,
  n.iter = 500,
  alpha = 0.1
)
}
\arguments{
\item{cor.list}{A matrix from Run_LINE()}

\item{main}{A character for the name of the group to compared to.}

\item{ref}{A character for the name of the group to compared to.}

\item{direction}{A character for the direction of enrichment}

\item{min.gs.size}{An integer indicating the minimum number of genes to used. Default is 20.}

\item{n.iter}{An integer indicating the number of background to generate for normlizing ES. Default is 500}

\item{alpha}{A numeric variable for gsea parameter}

\item{current.gs}{A character vector for a set of targets. One element from output of GetLE().}

\item{mir}{A character for the name of the miRNA to bet tested.}
}
\value{
A character vector of target names
}
\description{
Do backward random search to find set of targets that maximize the normalized dES
}
