% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.r
\name{Generate_simulation}
\alias{Generate_simulation}
\title{Generate Simulation Data}
\usage{
Generate_simulation(Ntargets, Nvar.percent, Nsample)
}
\arguments{
\item{Ntargets}{An integer giving the number of targets per miRNA}

\item{Nvar.percent}{A numeric value giving the percentage of targets to be variable between groups}

\item{Nsample}{An integer indicating the number of samples per group}
}
\value{
A list containing simulated data
}
\description{
Generate Simulation Data
}
