% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetLE}
\alias{GetLE}
\title{Generate a list of LE genes to be prioritized}
\usage{
GetLE(
  cor.list,
  target.mat,
  miR,
  direction = c("Negative", "Positive"),
  alpha = 0.1,
  with.weight = TRUE
)
}
\arguments{
\item{cor.list}{A list of correlation matrices with miRNA on columns and mRNA on rows}

\item{target.mat}{A binary matrix indicating target with miRNA on columns and mRNA on rows}

\item{miR}{miRNA to plot}

\item{direction}{A character string specifying the direction of enrichment, must be "Negative", "Positive" or "Both"}

\item{alpha}{Weight for running AD/KS test. Default is 0.1}

\item{with.weight}{Logic of whether to return weights as well}
}
\value{
An list of leading edge genes for each group
}
\description{
Generate a list of LE genes to be prioritized
}
