% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffreg.R
\name{Pairwise_postpoc}
\alias{Pairwise_postpoc}
\title{Do post-hoc pair-wise ks tests for the significant diff-reg miRNAs}
\usage{
Pairwise_postpoc(
  diffreg.res,
  cor.list,
  target.mat,
  significance = c("Pval", "Perm.pval", "FDR"),
  cutoff = 0.05,
  alpha = NULL
)
}
\arguments{
\item{diffreg.res}{Differential regulation results from Run_diffreg()}

\item{cor.list}{A list of correlation matrices with miRNA on columns and mRNA on rows}

\item{target.mat}{A binary matrix indicating target with miRNA on columns and mRNA on rows}

\item{significance}{A character value for type of significance threshold}

\item{cutoff}{A numeric value for significance threshold}

\item{alpha}{Weight for running AD/KS test}
}
\value{
A list with ks scores and p-values matrices for each pair-wise comparison
}
\description{
Do post-hoc pair-wise ks tests for the significant diff-reg miRNAs
}
