% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffreg.R
\name{Plot_enrichment}
\alias{Plot_enrichment}
\title{Do post-hoc pair-wise ks tests for the significant diff-reg miRNAs}
\usage{
Plot_enrichment(
  cor.list,
  miR,
  target.mat = NULL,
  target.set = NULL,
  direction = "Negative",
  pal = NULL,
  alpha = NULL
)
}
\arguments{
\item{cor.list}{A list of correlation matrices with miRNA on columns and mRNA on rows}

\item{miR}{miRNA to plot}

\item{target.mat}{A binary matrix indicating target with miRNA on columns and mRNA on rows}

\item{target.set}{A vector of targets for plotting}

\item{direction}{A character string specifying the direction of enrichment, must be "Negative" or "Positive"}

\item{pal}{Pallette to use; default will be used if NULL}
}
\value{
An enrichment plot
}
\description{
Do post-hoc pair-wise ks tests for the significant diff-reg miRNAs
}
