% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffreg.R
\name{Run_diffreg}
\alias{Run_diffreg}
\title{Given a list of correlation matrcies, return the SpQN normalized matrices}
\usage{
Run_diffreg(
  cor.list,
  target.mat,
  min.target.number = 5,
  direction = "Negative",
  alpha = 0.1
)
}
\arguments{
\item{cor.list}{A list of correlation matrices with miRNA on columns and mRNA on rows}

\item{target.mat}{A binary matrix indicating target with miRNA on columns and mRNA on rows}

\item{min.target.number}{A integer indicating minimum number of targets requried for doing diffreg}

\item{direction}{A character string specifying the direction of enrichment, must be "Negative" or "Positive"}

\item{alpha}{Weight for running AD/KS test}
}
\value{
A data frame with ES and test results for each miRNA
}
\description{
Given a list of correlation matrcies, return the SpQN normalized matrices
}
