% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffreg.R
\name{Run_permutation}
\alias{Run_permutation}
\title{Given a diff-reg results, do permutation on group label}
\usage{
Run_permutation(
  diffreg.res,
  mrna.mat,
  mirna.mat,
  raw.mrna.mat = NULL,
  raw.mirna.mat = NULL,
  do.spqn = TRUE,
  target.mat,
  group.label,
  do.parallel = FALSE,
  n.iter = 1000,
  seed = 1234
)
}
\arguments{
\item{diffreg.res}{Differential regulation results from Run_diffreg()}

\item{mrna.mat}{A matrix of mRNA expression (genes at rows and samples at columns)}

\item{mirna.mat}{A matrix of miRNA expression (genes at rows and samples at columns)}

\item{raw.mrna.mat}{(optional) A raw matrix of mRNA expression (genes at rows and samples at columns)}

\item{raw.mirna.mat}{(optional) A raw matrix of miRNA expression (genes at rows and samples at columns)}

\item{do.spqn}{A logical variable indication whether SpQN should be done}

\item{target.mat}{A binary matrix indicating target with miRNA on columns and mRNA on rows}

\item{group.label}{A factor vector subtype label for the samples}

\item{do.parallel}{A logical variable indicating whether parallization will be performed}

\item{n.iter}{An integer indicating the number of interaions for n.iter}

\item{seed}{A integer for permutation seed}
}
\value{
A data frame similar to output from Run_diffreg() with additional permutation significance values
}
\description{
Given a diff-reg results, do permutation on group label
}
