% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSpQN.R
\name{RunGroupSpQN}
\alias{RunGroupSpQN}
\title{Given a list of correlation matrcies, return the SpQN normalized matrices}
\usage{
RunGroupSpQN(mrna.mat, mirna.mat, group.label, cor.list, numCores = 1, ...)
}
\arguments{
\item{mrna.mat}{An expression matrix for mRNAs}

\item{mirna.mat}{An expression matrix for miRNAs}

\item{group.label}{A factor object for subtype/group}

\item{cor.list}{A list object from BuildCorList()}

\item{numCores}{An integer indicating the number of cores to use}

\item{...}{Arguments passed to normalize_correlation()}
}
\value{
A list of normalized correlation matrices
}
\description{
Given a list of correlation matrcies, return the SpQN normalized matrices
}
