% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri_target_graph.R
\name{selectK}
\alias{selectK}
\title{Try to find optimal cluster number fro fuzzy k meansand make plots for k selection}
\usage{
selectK(normalize_matrix, k.range = 2:10, do.plot = TRUE)
}
\arguments{
\item{normalize_matrix}{A matrix from Run_LINE()}

\item{k.range}{A numeric vector for k numbers to try}

\item{do.plot}{A Logic variable indicating whether plots are shown}
}
\value{
An enrichment plot
}
\description{
Try to find optimal cluster number fro fuzzy k meansand make plots for k selection
}
