% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated differential regulated data}
\format{
A list with four elements:
\describe{
\item{predict.target}{The predicted targets of each miRNAs}
\item{group.label}{Group labels for each individuals}
\item{mrna.mat}{Gene expression matrix}
\item{mirna.mat}{miRNA expression matrix}
...
}
}
\usage{
sim.data
}
\description{
A dataset containing simulated data for 300 samples with 1000 genes
and 20 miRNAs. The first 10 are true positive.
}
\keyword{datasets}
