% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildTargetMat}
\alias{BuildTargetMat}
\title{Generate a list of correlation matrices for later analysis}
\usage{
BuildTargetMat(
  cor.list,
  Pair.df = c(),
  gene.name = c("hgnc_symbol", "ensembl_ID")
)
}
\arguments{
\item{mrna.mat}{A matrix of mRNA expression (genes at rows and samples at columns)}

\item{mirna.mat}{A matrix of miRNA expression (genes at rows and samples at columns)}
}
\value{
A binary target matrix with mRNA on rows and miRNA on cols
}
\description{
Generate a list of correlation matrices for later analysis
}
