% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSpQN.R
\name{plot_signal_condition_exp}
\alias{plot_signal_condition_exp}
\title{Plot density plot showing signal (most negative correlation) vs. background per bin}
\usage{
plot_signal_condition_exp(cor_mat, ave_exp1, ave_exp2, signal = 0.001)
}
\description{
Plot density plot showing signal (most negative correlation) vs. background per bin
}
